Merged PR 1710: Infrastructure to allow testing in build jobs that will run against PRs in future.
Dockerfile:
Run the docker container as non-root, either as the USERID:GROUPID provided in build-args or 1000:1000 by default.

settings.py:
Add XML reporting on the django tests.
Remove redis: use memcaching for the test container.

docker-compose.yml:
Don't mount the directory as this doesn't work well if this is docker inside docker (it'll try and mount the host and find no code).
Set some dummy ENV vars so we don't need a secrets.env file.

Add a build.py file that VSTS can call for running tests.
Add a setup_python.py that sets up a virtualenv (in this case not actually needed by VSTS, as we test in docker).

Add a .dockerignore file for any test files and compiled python lying around.
Tweak .gitignore.

This depends on the build_image updates to work.

Related work items: #3010