Merged PR 1755: Create an nginx image for serving the Django media files.
Specify a very simple, standalone nginx Docker image that will serve the boticon and any other images that the console needs via the MEDIA_URL.

Although the mediaserver is not technically a dependency of Django (any server that can honour the MEDIA_URL will do), it makes sense to put it next to Django rather than at the system level. Also it is likely to not really grow beyond the single Dockerfile and nginx.conf file, but these will probably be tweaked over time.

The Docker config will make sure that this mediaserver image is created and that Django has the correct MEDIA_URL pointing to it.

Tested locally by hacking around with images in Docker volumes mounted RO in nginx, simulating the botstore and seeing that Django serves these.

Add a bit of structure to this repo, as the source code was already getting mixed up with other non-essential files even before the need to add mediaserver:
1) Move all existing Django code under src/
2) Create mediaserver to contain the Dockerfile and nginx.conf file

Tweaked some Docker GIT files (.dockerignore, .gitignore)

Related work items: #3070