Merged PR 1860: Basic Facebook integration
The integration tab on the console allows the user to connect to a Facebook account and integrate this bot to Facebook.
This change allows the user to connect to and disconnect from Facebook, to get a list of the user's Facebook pages and integrate with one of them.

Note that to test the connect your browser has got to be pointed to dev.hutoma.com or fb.hutoma.com
Easiest way to do this is to add fb.hutoma.com to your hosts file and set it to 127.0.0.1 or the address of your docker host.

Related work items: #3058, #3059